requirements.txt
Django~=2.1.4
django-filter~=2.0.0
djangorestframework~=3.9.0
